requirements.txt
fastapi==0.100.0
hypercorn==0.14.4
psycopg2-binary==2.9.11
python-dotenv==1.0.1
pydantic==2.12.4
